update
Minor edits
