+++ src/lang/en/drivers.json
@@ -363,7 +363,6 @@
     "password": "Password",
     "phone": "Phone",
     "root_folder_id": "Root folder id",
-    "root_folder_id-tips": "be careful when using the -11 value, some operations may cause system errors",
     "upload_thread": "Upload thread",
     "upload_thread-tips": "1\u003c=thread\u003c=32"
   },
